# apps/api/src/modules/xac-thuc/dto/yeu-cau-dat-lai-mat-khau.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEmail, MaxLength } from 'class-validator';

export class YeuCauDatLaiMatKhauDto {
  @ApiProperty({ example: 'khachhang@example.com' })
  @IsEmail()
  @MaxLength(191)
  email!: string;
}